+++ Projects/Editor/Editor.cbp
@@ -113,6 +113,8 @@ Allow attacking to occur in 360 directions (not just 4)]]>
 		<Unit filename="../../Source/include/Shared/GUI/Form.hpp" />
 		<Unit filename="../../Source/include/Shared/GUI/ResizableImage.hpp" />
 		<Unit filename="../../Source/include/Shared/GUI/WrappingBox.hpp" />
+		<Unit filename="../../Source/include/Shared/Items/Item.hpp" />
+		<Unit filename="../../Source/include/Shared/Items/ItemEffect.hpp" />
 		<Unit filename="../../Source/include/Shared/Maps/Map.hpp" />
 		<Unit filename="../../Source/include/Shared/Maps/MapManager.hpp" />
 		<Unit filename="../../Source/include/Shared/Maps/Map_priv.hpp" />
@@ -121,8 +123,6 @@ Allow attacking to occur in 360 directions (not just 4)]]>
 		<Unit filename="../../Source/include/Shared/Media/Animation.hpp" />
 		<Unit filename="../../Source/include/Shared/Media/Playlist.hpp" />
 		<Unit filename="../../Source/include/Shared/Media/SoundEngine.hpp" />
-		<Unit filename="../../Source/include/Shared/Objects/Item.hpp" />
-		<Unit filename="../../Source/include/Shared/Objects/ItemEffect.hpp" />
 		<Unit filename="../../Source/include/Shared/Particles/Particle.hpp" />
 		<Unit filename="../../Source/include/Shared/Particles/ParticleGenerator.hpp" />
 		<Unit filename="../../Source/include/Shared/Particles/ParticleGeneratorFactory.hpp" />
@@ -183,6 +183,8 @@ Allow attacking to occur in 360 directions (not just 4)]]>
 		<Unit filename="../../Source/src/Shared/GUI/CellTable.cpp" />
 		<Unit filename="../../Source/src/Shared/GUI/Form.cpp" />
 		<Unit filename="../../Source/src/Shared/GUI/WrappingBox.cpp" />
+		<Unit filename="../../Source/src/Shared/Items/Item.cpp" />
+		<Unit filename="../../Source/src/Shared/Items/ItemEffect.cpp" />
 		<Unit filename="../../Source/src/Shared/Maps/Map.cpp" />
 		<Unit filename="../../Source/src/Shared/Maps/MapManager.cpp" />
 		<Unit filename="../../Source/src/Shared/Maps/Tileset.cpp" />
@@ -190,8 +192,6 @@ Allow attacking to occur in 360 directions (not just 4)]]>
 		<Unit filename="../../Source/src/Shared/Media/Animation.cpp" />
 		<Unit filename="../../Source/src/Shared/Media/Playlist.cpp" />
 		<Unit filename="../../Source/src/Shared/Media/SoundEngine.cpp" />
-		<Unit filename="../../Source/src/Shared/Objects/Item.cpp" />
-		<Unit filename="../../Source/src/Shared/Objects/ItemEffect.cpp" />
 		<Unit filename="../../Source/src/Shared/Particles/Particle.cpp" />
 		<Unit filename="../../Source/src/Shared/Particles/ParticleGenerator.cpp" />
 		<Unit filename="../../Source/src/Shared/Particles/ParticleGeneratorFactory.cpp" />
