+++ Source/include/Editor/Helpers/MenuGenerators.hpp
@@ -2,7 +2,7 @@
 #define MENUGENERATORS_HPP
 
 #include "Shared/GUI/Form.hpp"
-#include "Shared/Objects/ItemEffect.hpp"
+#include "Shared/Items/ItemEffect.hpp"
 
 void addItemEffectsToForm(Form& form, const ItemEffect::List& effects);
 ItemEffect::List getItemEffectsFromForm(Form& form);
