+++ Source/include/Shared/Items/ItemEffect.hpp
